Enable Travis CI builds for the iOS version of AffdexMe
This should help with pull requests and give users a starting place for continuous integration with their apps.

I had to add a shared scheme to make this work because xctool, like xcodebuild, isn't able to automatically create schemes.  They suggested sharing a scheme and checking it into source control.